--- a/src/coreclr/inc/gcinfodecoder.h
+++ b/src/coreclr/inc/gcinfodecoder.h
@@ -218,7 +218,7 @@ enum GcInfoDecoderFlags
     DECODE_INTERRUPTIBILITY      = 0x08,
     DECODE_GC_LIFETIMES          = 0x10,
     DECODE_NO_VALIDATION         = 0x20,
-    DECODE_PSP_SYM               = 0x40,
+    DECODE_PSP_SYM               = 0x40,    // Unused starting with v4 format
     DECODE_GENERICS_INST_CONTEXT = 0x80,    // stack location of instantiation context for generics
                                             // (this may be either the 'this' ptr or the instantiation secret param)
     DECODE_GS_COOKIE             = 0x100,   // stack location of the GS cookie
@@ -237,7 +237,7 @@ enum GcInfoHeaderFlags
     GC_INFO_IS_VARARG                   = 0x1,
     // unused                           = 0x2, // was GC_INFO_HAS_SECURITY_OBJECT
     GC_INFO_HAS_GS_COOKIE               = 0x4,
-    GC_INFO_HAS_PSP_SYM                 = 0x8,
+    GC_INFO_HAS_PSP_SYM                 = 0x8, // Unused starting with v4 format
     GC_INFO_HAS_GENERICS_INST_CONTEXT_MASK   = 0x30,
     GC_INFO_HAS_GENERICS_INST_CONTEXT_NONE   = 0x00,
     GC_INFO_HAS_GENERICS_INST_CONTEXT_MT     = 0x10,