--- a/src/coreclr/jit/codegen.h
+++ b/src/coreclr/jit/codegen.h
@@ -462,8 +462,6 @@ protected:
         regMaskTP fiSaveRegs;                  // Set of registers saved in the funclet prolog (includes LR)
         unsigned  fiFunctionCallerSPtoFPdelta; // Delta between caller SP and the frame pointer
         unsigned  fiSpDelta;                   // Stack pointer delta
-        unsigned  fiPSP_slot_SP_offset;        // PSP slot offset from SP
-        int       fiPSP_slot_CallerSP_offset;  // PSP slot offset from Caller SP
     };
 
     FuncletFrameInfoDsc genFuncletInfo;
@@ -479,9 +477,7 @@ protected:
         int fiFunction_CallerSP_to_FP_delta; // Delta between caller SP and the frame pointer in the parent function
                                              // (negative)
         int fiSP_to_FPLR_save_delta;         // FP/LR register save offset from SP (positive)
-        int fiSP_to_PSP_slot_delta;          // PSP slot offset from SP (positive)
         int fiSP_to_CalleeSave_delta;        // First callee-saved register slot offset from SP (positive)
-        int fiCallerSP_to_PSP_slot_delta;    // PSP slot offset from Caller SP (negative)
         int fiFrameType;                     // Funclet frame types are numbered. See genFuncletProlog() for details.
         int fiSpDelta1;                      // Stack pointer delta 1 (negative)
         int fiSpDelta2;                      // Stack pointer delta 2 (negative)
@@ -498,7 +494,6 @@ protected:
     {
         unsigned fiFunction_InitialSP_to_FP_delta; // Delta between Initial-SP and the frame pointer
         unsigned fiSpDelta;                        // Stack pointer delta
-        int      fiPSP_slot_InitialSP_offset;      // PSP slot offset from Initial-SP
     };
 
     FuncletFrameInfoDsc genFuncletInfo;
@@ -514,8 +509,6 @@ protected:
         int fiFunction_CallerSP_to_FP_delta; // Delta between caller SP and the frame pointer in the parent function
                                              // (negative)
         int fiSP_to_CalleeSaved_delta;       // CalleeSaved register save offset from SP (positive)
-        int fiSP_to_PSP_slot_delta;          // PSP slot offset from SP (positive)
-        int fiCallerSP_to_PSP_slot_delta;    // PSP slot offset from Caller SP (negative)
         int fiSpDelta;                       // Stack pointer delta (negative)
     };
 
@@ -611,80 +604,6 @@ protected:
     void genFuncletProlog(BasicBlock* block);
     void genFuncletEpilog();
     void genCaptureFuncletPrologEpilogInfo();
-
-    /*-----------------------------------------------------------------------------
-     *
-     *  Set the main function PSPSym value in the frame.
-     *  Funclets use different code to load the PSP sym and save it in their frame.
-     *  See the document "CLR ABI.md" for a full description of the PSPSym.
-     *  The PSPSym section of that document is copied here.
-     *
-     ***********************************
-     *  The name PSPSym stands for Previous Stack Pointer Symbol.  It is how a funclet
-     *  accesses locals from the main function body.
-     *
-     *  First, two definitions.
-     *
-     *  Caller-SP is the value of the stack pointer in a function's caller before the call
-     *  instruction is executed. That is, when function A calls function B, Caller-SP for B
-     *  is the value of the stack pointer immediately before the call instruction in A
-     *  (calling B) was executed. Note that this definition holds for both AMD64, which
-     *  pushes the return value when a call instruction is executed, and for ARM, which
-     *  doesn't. For AMD64, Caller-SP is the address above the call return address.
-     *
-     *  Initial-SP is the initial value of the stack pointer after the fixed-size portion of
-     *  the frame has been allocated. That is, before any "alloca"-type allocations.
-     *
-     *  The PSPSym is a pointer-sized local variable in the frame of the main function and
-     *  of each funclet. The value stored in PSPSym is the value of Initial-SP/Caller-SP
-     *  for the main function.  The stack offset of the PSPSym is reported to the VM in the
-     *  GC information header.  The value reported in the GC information is the offset of the
-     *  PSPSym from Initial-SP/Caller-SP. (Note that both the value stored, and the way the
-     *  value is reported to the VM, differs between architectures. In particular, note that
-     *  most things in the GC information header are reported as offsets relative to Caller-SP,
-     *  but PSPSym on AMD64 is one (maybe the only) exception.)
-     *
-     *  The VM uses the PSPSym to find other locals it cares about (such as the generics context
-     *  in a funclet frame). The JIT uses it to re-establish the frame pointer register, so that
-     *  the frame pointer is the same value in a funclet as it is in the main function body.
-     *
-     *  When a funclet is called, it is passed the Establisher Frame Pointer. For AMD64 this is
-     *  true for all funclets and it is passed as the first argument in RCX, but for ARM this is
-     *  only true for first pass funclets (currently just filters) and it is passed as the second
-     *  argument in R1. The Establisher Frame Pointer is a stack pointer of an interesting "parent"
-     *  frame in the exception processing system. For the CLR, it points either to the main function
-     *  frame or a dynamically enclosing funclet frame from the same function, for the funclet being
-     *  invoked. The value of the Establisher Frame Pointer is Initial-SP on AMD64, Caller-SP on ARM.
-     *
-     *  Using the establisher frame, the funclet wants to load the value of the PSPSym. Since we
-     *  don't know if the Establisher Frame is from the main function or a funclet, we design the
-     *  main function and funclet frame layouts to place the PSPSym at an identical, small, constant
-     *  offset from the Establisher Frame in each case. (This is also required because we only report
-     *  a single offset to the PSPSym in the GC information, and that offset must be valid for the main
-     *  function and all of its funclets). Then, the funclet uses this known offset to compute the
-     *  PSPSym address and read its value. From this, it can compute the value of the frame pointer
-     *  (which is a constant offset from the PSPSym value) and set the frame register to be the same
-     *  as the parent function. Also, the funclet writes the value of the PSPSym to its own frame's
-     *  PSPSym. This "copying" of the PSPSym happens for every funclet invocation, in particular,
-     *  for every nested funclet invocation.
-     *
-     *  On ARM, for all second pass funclets (finally, fault, catch, and filter-handler) the VM
-     *  restores all non-volatile registers to their values within the parent frame. This includes
-     *  the frame register (R11). Thus, the PSPSym is not used to recompute the frame pointer register
-     *  in this case, though the PSPSym is copied to the funclet's frame, as for all funclets.
-     *
-     *  Catch, Filter, and Filter-handlers also get an Exception object (GC ref) as an argument
-     *  (REG_EXCEPTION_OBJECT).  On AMD64 it is the second argument and thus passed in RDX.  On
-     *  ARM this is the first argument and passed in R0.
-     *
-     *  (Note that the JIT64 source code contains a comment that says, "The current CLR doesn't always
-     *  pass the correct establisher frame to the funclet. Funclet may receive establisher frame of
-     *  funclet when expecting that of original routine." It indicates this is the reason that a PSPSym
-     *  is required in all funclets as well as the main function, whereas if the establisher frame was
-     *  correctly reported, the PSPSym could be omitted in some cases.)
-     ***********************************
-     */
-    void genSetPSPSym(regNumber initReg, bool* pInitRegZeroed);
 
     void genUpdateCurrentFunclet(BasicBlock* block);