… docs/design/coreclr/botr/clr-abi.md
… docs/design/coreclr/botr/guide-for-porting.md
… docs/design/features/OsrDetailsAndDebugging.md
… src/coreclr/jit/codegenarmarch.cpp
… src/coreclr/jit/codegencommon.cpp
… src/coreclr/jit/codegenriscv64.cpp
… src/coreclr/jit/codegenxarch.cpp
… src/coreclr/jit/compiler.h
… src/coreclr/jit/compiler.hpp
… src/coreclr/jit/emit.h
… src/coreclr/jit/emitarm64.cpp
… src/coreclr/jit/flowgraph.cpp
… src/coreclr/jit/gcencode.cpp
… src/coreclr/jit/gentree.cpp
… src/coreclr/jit/lclvars.cpp
… src/coreclr/jit/targetamd64.h
… src/coreclr/nativeaot/Runtime/amd64/ExceptionHandling.S
… src/coreclr/nativeaot/Runtime/amd64/ExceptionHandling.asm
… src/coreclr/nativeaot/Runtime/inc/ModuleHeaders.h
… src/coreclr/pal/inc/unixasmmacrosarm64.inc
… src/coreclr/tools/Common/Internal/Runtime/ModuleHeaders.cs
… src/coreclr/vm/amd64/AsmHelpers.asm
… src/coreclr/vm/amd64/asmhelpers.S
… src/coreclr/vm/amd64/profiler.cpp
… src/coreclr/vm/arm/ehhelpers.S
… src/coreclr/vm/arm/profiler.cpp
… src/coreclr/vm/arm64/asmhelpers.S
… src/coreclr/vm/arm64/asmhelpers.asm
… src/coreclr/vm/arm64/profiler.cpp
… src/coreclr/vm/codeman.h