@filipnavara 2025-04-14
Diffs look favorable:
https://dev.azure.com/dnceng-public/public/_build/results?buildId=1015738&view=ms.vss-build-web.run-extensions-tab

ARM64 has places where it stopped using vector registers for memory copying. Not sure if we can tweak the alignment to fix this but it seems plausible.

/cc @jkotas