@jkotas 2025-04-15
> ARM64 has places where it stopped using vector registers for memory copying

What's an example of a method with this diff?