@filipnavara 2025-04-16
# Benchmarks on win-x64

summary:
better: 18, geomean: 1.024
worse: 1, geomean: 1.020
total diff: 19

| Slower                                                            | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| ----------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| Exceptions.Handling.TryAndCatchDeep_CaugtAtTheTop(kind: Hardware) |      1.02 |          4627.33 |          4721.84 |         |

| Faster                                                                   | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ------------------------------------------------------------------------ | ---------:| ----------------:| ----------------:| --------:|
| Exceptions.Handling.TryAndFinallyDeep(kind: Software)                    |      1.10 |          4402.52 |          3989.19 |         |
| Exceptions.Handling.TryAndFinallyDeep(kind: Hardware)                    |      1.03 |          4866.53 |          4722.65 |         |
| Exceptions.Handling.MultipleNestedTryCatch_FirstCatches(kind: Hardware)  |      1.03 |          1865.55 |          1815.92 |         |
| Exceptions.Handling.CatchAndRethrowDeep(kind: Software)                  |      1.03 |         13680.62 |         13316.77 |         |
| Exceptions.Handling.MultipleNestedTryFinally(kind: Hardware)             |      1.03 |          1957.12 |          1907.28 |         |
| Exceptions.Handling.CatchAndRethrowDeep(kind: Hardware)                  |      1.03 |         14534.44 |         14178.08 |         |
| Exceptions.Handling.ThrowAndCatchWhenFinally(kind: Software)             |      1.02 |          1146.96 |          1120.37 |         |
| Exceptions.Handling.ThrowAndCatchDeep(kind: Software)                    |      1.02 |          3453.94 |          3389.16 |         |
| Exceptions.Handling.MultipleNestedTryFinally(kind: Software)             |      1.02 |          1201.91 |          1179.72 |         |
| Exceptions.Handling.ThrowAndCatchWhen(kind: Software)                    |      1.02 |          1140.83 |          1120.00 |         |
| Exceptions.Handling.ThrowAndCatchDeep(kind: ReflectionHardware)          |      1.02 |          6684.12 |          6572.26 |         |
| Exceptions.Handling.CatchAndThrowOtherDeep(kind: Hardware)               |      1.02 |         16761.70 |         16486.51 |         |
| Exceptions.Handling.CatchAndThrowOtherDeep(kind: Software)               |      1.02 |         16003.93 |         15744.48 |         |
| Exceptions.Handling.ThrowAndCatch(kind: ReflectionSoftware)              |      1.02 |          3481.91 |          3425.56 |         |
| Exceptions.Handling.ThrowAndCatchDeepRecursive(kind: ReflectionHardware) |      1.02 |          6896.87 |          6793.30 |         |
| Exceptions.Handling.ThrowAndCatchWhen(kind: Hardware)                    |      1.01 |          1887.48 |          1859.73 |         |
| Exceptions.Handling.ThrowAndCatchWhenFinally(kind: Hardware)             |      1.01 |          1894.94 |          1867.16 |         |
| Exceptions.Handling.ThrowAndCatchFinally(kind: Software)                 |      1.01 |          1094.52 |          1081.85 |         |